% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_info}
\alias{get_info}
\title{Lookup information about a function by name}
\usage{
get_info(fun.name)
}
\arguments{
\item{fun.name}{Function name to lookup}
}
\description{
Lookup information about a function by name
}

