% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/optim-functions.R
\docType{package}
\name{optim-functions}
\alias{optim-functions}
\alias{optim-functions-package}
\title{optim-functions: A collection of standard optimization functions along with
                 a standard interface to call and sample those functions.}
\description{
optim-functions: A collection of standard optimization functions along with
                 a standard interface to call and sample those functions.
}
\section{Sampling functions}{

The sampling functions all need to have a standard interface. They all must
take 2 parameters: \code{n}, the number of samples to generate and \code{k},
the number of dimensions to sample. They all return a matrix-like object
containing \code{n} rows of \code{k} columns of samples.
}

