% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{samplingf}
\alias{cartesian.sample}
\alias{halton.sample}
\alias{hexagonal.sample}
\alias{lh.sample}
\alias{random.sample}
\alias{samplingf}
\alias{sobol.sample}
\alias{torus.sample}
\title{The sampling functions take (n, k) where n is the number of samples
and k is the number of dimensions. The sampling functions should return
samples in a 0-1 hypercube.
sampling.func <- function(n, k)}
\usage{
lh.sample(n, k, ...)

cartesian.sample(n, k, ...)

random.sample(n, k, ...)

hexagonal.sample(n, k, ...)

torus.sample(n, k, ...)

sobol.sample(n, k, ...)

halton.sample(n, k, ...)
}
\arguments{
\item{n}{The number of samples.}

\item{k}{The number of dimensions.}

\item{...}{Additional arguments for the sampling (e.g. \code{seed}).}
}
\description{
The sampling functions take (n, k) where n is the number of samples
and k is the number of dimensions. The sampling functions should return
samples in a 0-1 hypercube.
sampling.func <- function(n, k)
}

